requirements.txt
groq
pydantic
python-dotenv
fastapi
uvicorn
typing
sentence_transformers
sklearn
numpy